requirements.txt
Pillow
Numpy
pyyaml